auto refresh button implemented
